+++ kernel/09_memory/libc/mem.h
@@ -3,7 +3,7 @@
 
 #include "../cpu/types.h"
 
-void memory_copy(uint8_t *source, uint8_t *dest, int nbytes);
+void memory_copy(uint8_t *source, uint8_t *dest, size_t nbytes);
 void memory_set(uint8_t *dest, uint8_t val, uint32_t len);
 
 #define PAGE_SIZE 4096 // 4KiB page size
